requirements.txt
discord.py
contextlib
async-timeout